Route composed scenario authority through owner admissions
